+++ src/primality.rs
@@ -1,6 +1,6 @@
 
 pub mod primality {
-    pub fn is_prime(n: u32) -> bool {
+    pub fn is_prime(n: u64) -> bool {
         let mut a = 1;
         let largest = loop {
             if a*a > n {
@@ -9,7 +9,7 @@ pub mod primality {
                 a += 1;
             }
         };
-        let largest_num_tested: u32 = largest;
+        let largest_num_tested: u64 = largest;
 
         for num in 2..largest_num_tested {
             if n % num == 0 {
